Support standard document types
Ideally, this site will be implemented using e.g. ASP.NET Core MVC or Razor Pages. But in the meantime, we want to ensure that static pages can be utilized to get the ball rolling. This allows e.g. `Index.html` to be used for that purpose, if no other route is present.